moved heatmap logic into separate file
