+++ stregsystem/views.py
@@ -53,6 +53,8 @@ from .forms import MobilePayToolForm, QRPaymentForm, PurchaseForm, RankingDateFo
 
 import json
 
+from .purchase_heatmap import __get_purchase_heatmap_data
+
 
 def __get_news():
     try:
@@ -83,90 +85,6 @@ def __organize_purchase_heatmap_data(heatmap_data: list, start_date: datetime.da
                 new_list[i].append(heatmap_data[j])
 
     return new_list
-
-
-def __get_purchase_heatmap_day_color(
-    products: List[Product],
-    products_by_color: (
-        List[Product],
-        List[Product],
-        List[Product],
-    ),
-) -> (int, int, int,):
-    category_representation = [0, 0, 0]
-
-    for product in products:
-        for category_products_index in range(3):
-            if product in products_by_color[category_products_index]:
-                category_representation[category_products_index] = category_representation[category_products_index] + 1
-
-    total_category_sum = sum(category_representation)
-
-    if total_category_sum == 0:
-        return 235, 237, 240  # Grey
-
-    return tuple(category_sum / total_category_sum * 255 for category_sum in category_representation)
-
-
-def __get_purchase_heatmap_data(
-    member: Member,
-    end_date: datetime.datetime,
-    weeks_to_display: int,
-    category_name_color: (
-        str,
-        str,
-        str,
-    ),
-) -> list:
-    from datetime import timedelta
-
-    days_to_go_back = (7 * weeks_to_display) - (6 - end_date.weekday() - 1)
-    cutoff_date = end_date.date() - timedelta(days=days_to_go_back)
-    last_sale_list = iter(
-        member.sale_set.filter(timestamp__gte=cutoff_date, timestamp__lte=end_date).order_by('-timestamp')
-    )
-
-    products_by_day = []
-    dates_by_day = []
-
-    try:
-        next_sale = next(last_sale_list)
-        next_sale_date = next_sale.timestamp.date()
-    except StopIteration:
-        next_sale = None
-        next_sale_date = None
-        pass
-
-    for single_date in (end_date - timedelta(days=n) for n in range(days_to_go_back)):
-        products_by_day.append([])
-        dates_by_day.append(single_date.date())
-
-        try:
-            while next_sale_date == single_date.date():
-                products_by_day[-1].append(next_sale.product)
-
-                next_sale = next(last_sale_list)
-                next_sale_date = next_sale.timestamp.date()
-        except StopIteration:
-            next_sale = None
-            next_sale_date = None
-            continue
-
-    category_by_name = [Category.objects.filter(name=cat_name) for cat_name in category_name_color]
-    products_by_category = tuple(
-        Product.objects.filter(categories__in=category_ins) for category_ins in category_by_name
-    )
-
-    # Proposed format:
-    # Returned list: [<<Day 0 (today)>>, <<Day 1 (yesterday)>>, ...]
-    # Day item: (<<Date>>, <<RGB values tuple-format>>, [<<item ID 1>>, ...])
-    days = []
-
-    for day_index in range(len(products_by_day)):
-        day_color = __get_purchase_heatmap_day_color(products_by_day[day_index], products_by_category)
-        days.append((dates_by_day[day_index], day_color, [x.id for x in products_by_day[day_index]]))
-
-    return days
 
 
 def roomindex(request):
@@ -315,14 +233,13 @@ def usermenu(request, room, member, bought, from_sale=False):
 
     # heatmap logic begin
     weeks_to_display = 10
-    from datetime import datetime
 
     raw_heatmap_data = __get_purchase_heatmap_data(
-        member, datetime.today(), weeks_to_display, ("beer", "energy", "soda")
+        member, datetime.datetime.today(), weeks_to_display, ("beer", "energy", "soda")
     )
-    heatmap_data = __organize_purchase_heatmap_data(raw_heatmap_data[::-1], datetime.today())
+    heatmap_data = __organize_purchase_heatmap_data(raw_heatmap_data[::-1], datetime.datetime.today())
     row_labels = ["", "Mandag", "", "Onsdag", "", "Fredag", ""]
-    current_week = datetime.today().isocalendar()[1]
+    current_week = datetime.datetime.today().isocalendar()[1]
     column_labels = [str(current_week - x) for x in range(weeks_to_display)][::-1]
 
     rows = zip(row_labels, heatmap_data)
